save for safety

+++ Gymnasiearbete/src/gameplay/Water.cpp
@@ -16,7 +16,7 @@ void Water::Update(float deltaTime)
     memset(m_Loaded, 0, sizeof(m_Loaded[0][0]) * 256 * 256);
 
     glm::vec3 pos = Camera::GetFollowTarget()->m_Position;
-    glm::ivec2 chunkPos((int)(pos.x / 17.999f) + 128, (int)(pos.z / 17.999f) + 128);
+    glm::vec2 chunkPos((int)(pos.x / 17.999f) + 128, (int)(pos.z / 17.999f) + 128);
 
     for (int x = chunkPos.x - 4; x < chunkPos.x + 4; x++)
     {
